pyproject.toml
[tool.poetry]
name = "django-florist-app"
version = "0.1.0"
description = ""
authors = ["ivan <shanin-i2011@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.*"
django-environ = "^0.11.2"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.27.0 "
django-redis = "^5.4.0"
django = "^4.2.8"
psycopg = { extras = ["c"], version = "^3.1.14" }
django-autoslug = "^1.9.9"
pillow = "^10.1.0"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
django-cachalot = "^2.6.2"
dj-rest-auth = "^5.0.2"
djangorestframework-simplejwt = "^5.3.1"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
isort = "^5.13.0"
flake8-pyproject = "^1.2.3"
pre-commit = "^3.6.0"
pytest-django = "^4.7.0"
pytest-factoryboy = "^2.6.0"
django-debug-toolbar = "^4.2.0"

[tool.flake8]
max-line-length = 79
exclude = [".tox", ".git", "*/migrations/*", "*/static/CACHE/*", "docs", "venv", "settings.py"]

[tool.black]
line-length = 79
target-version = ['py311']
force-exclude = 'settings\.py'


[tool.isort]
profile = "black"
line_length = 79
skip = ["venv/"]
skip_glob = ["**/migrations/*.py"]


[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "src.config.settings"
python_files = [
    "tests.py",
    "test_*.py",
]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
